+++ NEWS.md
@@ -1,3 +1,8 @@
+# nemoRsem 0.2.0.0 
+<span style="color:grey;">16/05/2023</span>
+
+* `scheme_interp_slice` now returns a scheme to extract an array layer if the target depth already exists in the dataset.
+
 # nemoRsem 0.1.0.0 
 <span style="color:grey;">16/05/2023</span>